update.bash: use kill -s SIGHUP instead of systemctl stop
We want to reload the sequence file, but we don't necessarily need to stop and
restart the entire service. Further, if we stop and restart the service,
entries added to the IP Table will remain if this happens before the timeout
for delete. This may be a bug in knockd itself, but it's still good practice to
use SIGHUP in this case instead of restarting the whole service.